pyproject.toml
[project]
name = "azure-ai-agent-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-projects>=1.0.0b4",
    "azure-identity>=1.19.0",
    "gradio>=5.9.1",
    "python-dotenv>=1.0.1",
]
